+++ packages/component-library/draft/Kaizen/ZenOffCanvas/ZenOffCanvas.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { ColorScheme } from "@kaizen/component-library/draft/Kaizen/ZenNavigationBar"
+import { ColorScheme } from "@kaizen/draft-zen-navigation-bar/KaizenDraft/ZenNavigationBar/NavigationBar"
 import classNames from "classnames"
 import Header from "./components/Header"
 import Menu from "./components/Menu"
@@ -73,7 +73,11 @@ export class ZenOffCanvas extends React.Component<Props> {
               <nav className={styles.links}>
                 {links &&
                   Object.keys(links).map(section => (
-                    <Menu section={section} link={links[section]} />
+                    <Menu
+                      key={section}
+                      section={section}
+                      link={links[section]}
+                    />
                   ))}
               </nav>
             </div>